Se agrega basic auth para generar el token de acceso y se agrega JWT a path de Usuarios

+++ routes/usuarios.py
@@ -2,12 +2,11 @@ from fastapi import APIRouter, Body, Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.orm import selectinload
 from db.models import TipoUsuario, Usuario
-from shared.dependencies import AsyncDbSessionDep, SyncDbSessionDep, get_user_from_db
+from shared.dependencies import AsyncDbSessionDep, SyncDbSessionDep, get_user_from_token_data
 from shared.schemas import UsuarioCreate, UsuarioOut
-from shared.utils import get_example
+from shared.utils import get_example, get_password_hash
 
-router = APIRouter(prefix="/usuarios", tags=["Usuarios"],
-                   dependencies=[Depends(get_user_from_db)])
+router = APIRouter(prefix="/usuarios", tags=["Usuarios"])
 
 
 @router.get(
@@ -17,8 +16,16 @@ router = APIRouter(prefix="/usuarios", tags=["Usuarios"],
 )
 async def read_users(
     db: AsyncDbSessionDep,
+    user: UsuarioOut | None = Depends(get_user_from_token_data),
 ):
-    """ Devuelve una lista con todos los usuarios. """
+    """
+    Devuelve una lista de todos los usuarios.
+
+    Requiere permisos de SMA para acceder a este recurso.
+    """
+    if user.tipo_usuario.tipo_usuario != "SMA":
+        raise HTTPException(status_code=403, detail="No tiene permisos para acceder a este recurso")
+    
     result = await db.scalars(select(Usuario).options(selectinload(Usuario.tipo_usuario)))
     users = result.all()
     return users
@@ -32,12 +39,15 @@ async def read_users(
 def read_user(
     id_usuario: int,
     db: SyncDbSessionDep,
+    user: UsuarioOut | None = Depends(get_user_from_token_data),
 ):
     """ 
     Devuelve un usuario por su id.
-    Argumentos: 
-    - id usuario (int)
+
+    Requiere permisos de SMA para acceder a este recurso.
     """
+    if user.tipo_usuario.tipo_usuario != "SMA":
+        raise HTTPException(status_code=403, detail="No tiene permisos para acceder a este recurso")
     usuario = db.query(Usuario).filter(Usuario.id_tipo_usuario == id_usuario).first()
     if not usuario:
         raise HTTPException(status_code=404, detail="No existe usuario con ese id")
@@ -49,6 +59,7 @@ def read_user(
 )
 def add_organismo(
     db: SyncDbSessionDep,
+    user: UsuarioOut | None = Depends(get_user_from_token_data),
     usuario: UsuarioCreate = Body(
         openapi_examples={
             "default": get_example("usuario_post"),
@@ -65,7 +76,12 @@ def add_organismo(
     - id tipo de usuario (int)
 
     Devuelve mensaje de confirmación con el recurso creado.
+
+    Requiere permisos de SMA para acceder a este recurso.
     """
+    if user.tipo_usuario.tipo_usuario != "SMA":
+        raise HTTPException(status_code=403, detail="No tiene permisos para acceder a este recurso")
+    
     if db.query(Usuario).filter(Usuario.email.ilike(usuario.email)).first():
         raise HTTPException(status_code=409, detail="Usuario ya existe")
     if (
@@ -81,6 +97,7 @@ def add_organismo(
         email=usuario.email,
         activo=usuario.activo,
         id_tipo_usuario=usuario.id_tipo_usuario,
+        password=get_password_hash(usuario.password),
     )
 
     db.add(data)
@@ -96,6 +113,7 @@ def add_organismo(
 def delete_usuario(
     id_usuario: int,
     db: SyncDbSessionDep,
+    user: UsuarioOut | None = Depends(get_user_from_token_data),
 ):
     """
     Elimina un usuario por su id.
@@ -103,7 +121,12 @@ def delete_usuario(
     - id usuario (int)
 
     Devuelve mensaje de confirmación.
+
+    Requiere permisos de SMA para acceder a este recurso.
     """
+    if user.tipo_usuario.tipo_usuario != "SMA":
+        raise HTTPException(status_code=403, detail="No tiene permisos para acceder a este recurso")
+    
     usuario = db.query(Usuario).filter(Usuario.id_usuario == id_usuario).first()
     if usuario:
         db.delete(usuario)
